requirements-test.txt
tox
nose
coverage
pyvirtualdisplay
pillow
coveralls
# osx error before pygame 2.0.0
pygame>=2.0.0.dev6
path.py
python-xlib
mss ; python_version > '3.4'
